+++ ChangeLog.md
@@ -3,6 +3,8 @@ MongoDB for XP Framework ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Added support for timestamp types - @thekid
+
 ## 0.1.0 / 2020-08-04
 
 * First public release - (@thekid)